Initial multi-part support (#1047)
Very initial multi-part support in the front end

1. The parts pane lets you pickparts
2. Selection is moved to per-part not just per-instrument
3. On change, select, etc... the front end messages to back
   end correctly

Basically the 'part-as-group' feature seems to work here.

Next up some of the actual part features like midi routing and
so on get checked.

+++ src/sample/multisample_support/multisample_import.cpp
@@ -82,7 +82,7 @@ bool importMultisample(const fs::path &p, engine::Engine &engine)
         return false;
     }
 
-    auto pt = std::clamp(engine.getSelectionManager()->selectedPart, 0, (int)numParts);
+    auto pt = std::clamp(engine.getSelectionManager()->selectedPart, (int16_t)0, (int16_t)numParts);
     auto &part = engine.getPatch()->getPart(pt);
 
     std::vector<int> addedGroupIndices;
